DRY: common functions in cli.binary, cli.registry and cli.remote
