+++ pacco/cli/registry.py
@@ -1,50 +1,10 @@
 import argparse
-import inspect
 import re
-from typing import Callable, Dict
 
-from pacco.cli import utils
-from pacco.cli.output_stream import OutputStream
-from pacco.manager.remote_manager import RemoteManager
+from pacco.cli.command_abstract import CommandAbstract
 
 
-class Registry:
-    def __init__(self, output: OutputStream, remote_manager: RemoteManager):
-        self.__out = output
-        self.__rm = remote_manager
-
-    def run(self, *args):
-        """
-        Entry point for executing commands, dispatcher to class methods.
-        """
-        if not args:
-            self.__show_help()
-            return
-        command = args[0]
-        remaining_args = args[1:]
-        commands = self.__get_commands()
-        if command not in commands:
-            if command in ["-h", "--help"]:
-                self.__show_help()
-                return
-            self.__out.writeln(
-                "'pacco registry {}' is an invalid command. See 'pacco registry --help'.".format(command),
-                error=True)
-            return
-        method = commands[command]
-        method(*remaining_args)
-
-    def __get_commands(self) -> Dict[str, Callable]:
-        result = {}
-        for method_name, method in inspect.getmembers(self, predicate=inspect.ismethod):
-            if not method_name.startswith('_') and method_name not in ["run"]:
-                result[method_name] = method
-        return result
-
-    def __show_help(self):
-        commands = self.__get_commands()
-        utils.show_help(self.__get_commands(), 'registry', self.__out)
-
+class Registry(CommandAbstract):
     def list(self, *args):
         """
         List registries of a remote.
